Changed floats to ints

+++ src/adts/Line.java
@@ -7,22 +7,22 @@ public class Line {
     /**
      * The first x coordinate
      */
-    private final float x1;
+    private final int x1;
     
     /**
      * The first y coordinate
      */
-    private final float y1;
+    private final int y1;
     
     /**
      * The second x coordinate
      */
-    private final float x2;
+    private final int x2;
     
     /**
      * The second y coordinate
      */
-    private final float y2;
+    private final int y2;
     
     /**
      * The stroke thickness
@@ -32,22 +32,22 @@ public class Line {
     /**
      * The alpha value
      */
-    private final float a;
+    private final int a;
     
     /**
      * The red value
      */
-    private final float r;
+    private final int r;
 
     /**
      * The green value
      */
-    private final float g;
+    private final int g;
     
     /**
      * The blue value
      */
-    private final float b;
+    private final int b;
     
     /**
      * @param x1 The first x coordinate
@@ -60,7 +60,7 @@ public class Line {
      * @param g The green value
      * @param b The blue value
      */
-    public Line(float x1, float y1, float x2, float y2, float strokeThickness, float a, float r, float g, float b){
+    public Line(int x1, int y1, int x2, int y2, float strokeThickness, int a, int r, int g, int b){
         this.x1 = x1;
         this.y1 = y1;
         this.x2 = x2;
@@ -79,28 +79,28 @@ public class Line {
     /**
      * @return The first x coordinate
      */
-    public float getX1() {
+    public int getX1() {
         return x1;
     }
 
     /**
      * @return The first y coordinate
      */
-    public float getY1() {
+    public int getY1() {
         return y1;
     }
 
     /**
      * @return The second x coordinate
      */
-    public float getX2() {
+    public int getX2() {
         return x2;
     }
 
     /**
      * @return The second y coordinate
      */
-    public float getY2() {
+    public int getY2() {
         return y2;
     }
 
@@ -114,34 +114,34 @@ public class Line {
     /**
      * @return The alpha value
      */
-    public float getA() {
+    public int getA() {
         return a;
     }
 
     /**
      * @return The red value
      */
-    public float getR() {
+    public int getR() {
         return r;
     }
 
     /**
      * @return The green value
      */
-    public float getG() {
+    public int getG() {
         return g;
     }
 
     /**
      * @return The blue value
      */
-    public float getB() {
+    public int getB() {
         return b;
     }
     
     @Override
     public String toString() {
-        return String.format("%f %f %f %f %f %f %f %f %f",this.getX1(),
+        return String.format("%d %d %d %d %f %d %d %d %d",this.getX1(),
                 this.getY1(),
                 this.getX2(),
                 this.getY2(),
